refactor: resolve pre-release hexarch follow-ups
