+++ internal/adapters/in/http/httphelper/localhost.go
@@ -1,4 +1,4 @@
-package httputil
+package httphelper
 
 import (
 	"net"
